Add sample Javadoc for Attendee.java
I have written up some sample Javadoc for the
Attendee class. It will be a good idea if we create
it as we code, since it is more descriptive, and that
way we have less work during reading week (we have
to provide Javadoc for our submission).

I also changed the name for two
methods, so that its more descriptive. I find it
better if the methods addSentMessages and
addReceivedMessages are instead called addSentMessage
and addReceivedMessage, respectively. This is because
we are only passing one Message as parameter at a
time. If we were passing a list, it would make more
sense to mak it plural.