Implement MessageManager.java
This commit adds the implementation of the
MessageManager and its corresponding tests in
MessageManagerTests.java. Currently, this
implementation saves the messages directly into the
user objects.

As it stands now, the Messenger Manager holds no fields,
just the required behaviour. If we decide to save the
messages directly in the MessageManager class, we must
agree on how we will we save them and how we would
initialize such class in the first place.