Add new functionality to UserManager.java
This commit fulfills the request by the
Controller-GatewayBy group. I have added a method that
returns a list of signed up users. One that returns a
list of the usernames for the users singed up for a given
event. For the date, time, and room of a given event, I
have created a method that returns them. Finally, I built
a method that returns all events with a given speaker.

Be careful when using these methods, they are not tested.
I will add new tests earlier tomorrow.