Update MessageManager and Message
This commit updates the MessageManager and Message
classes to account for the separation of the Use Case
classes.

The map functionality has been added to the
MessageManager, though it is not yet complete. The
construction of the map itself must still be
implemented, through the use of a gateway and a builder
class.

The message and broadcast methods now work without
using User and Event objects, just usernames. Thus, the
class is no longer dependent on those classes.