Fix getMessages in MessageManager.java
This commit fixes the getMessages method in the
MessageManager class. The method did not check whether
user had messages stored in the map, so when it looked
for those user's messages, it got a null pointer,
throwing a NullPointerException.

Now it checks whether the user has a message or not
beforehand.