Update MessageManager to remove messages
This commit adds new remove functionality to the
MessageManager. To remove a message, an actor using
this class must be able to enter a representation of
message without holding an actual reference to a
message (this to respect the Clean Architecture).

This commit also creates a new exception class used
in the manager.