Extract message storage from MessageManager
This commit extracts the storage functionality from the
MessageManager. This way, the MessageManager can become
independent from how the messages are stored.

A new interface, MessageCollection, contains all the
storage functionality. The MessageManager should be
injected with objects that follow this interface, and
these objects will take charge of storing the messages
in the manager.

This also makes it easier to create more than one message
collection. It is possible to create another message
collection for archived messages.

For this purpose, I have implemented the class MessageMap,
which implements the MessageCollection interface and uses
a hash map.

The NonExistentMessageException has also been renamed
NoSuchMessageException.