Add Filter Factory class
This commit adds a new filter factory under the Filter package. It is meant to
encapsulate the creation of the different types of filters, lowering dependence.

It also creates a default filter, the null filter, which does nothing. It also
adds a new Exception to handle wrong input in the EventFilterFactory.